// src/app/media/model/software-table.model.ts
export class SoftwareTable {
    softwareId: number;
    title: string;
    softwareOs: string;
    manufacturer: string;
    language: string;
    version: string;
    notice: string;
    linkValue: string;
    attachment: boolean;
    attachmentPath: string;
    attachmentName: string;
    attachmentType: string;
}